Separar Header y UserPanel
