+++ src/App.jsx
@@ -2,14 +2,25 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './style.css'
+import Header from './Header'
 import UserPanel from './UserPanel'
 
+const tiposUsuario = {
+  admin: ['Ver Reportes', 'Programacion', 'Configuración'],
+  cliente: ['Mis rutas', 'Historial', 'Soporte'],
+  invitado: ['Productos', 'Contacto']
+};
+
+
 function App() {
-  const [count, setCount] = useState(0)
+  const [count, setCount] = useState(0);
+  const [role, setRole] = useState(null);
+  const possible_roles = ["admin", "cliente", "invitado"];
 
   return (
     <>
-    <UserPanel/>
+    <Header setRole={setRole} possibleRoles={tiposUsuario}/>
+    <UserPanel tiposUsuario={tiposUsuario} selectedTipoUsu={role}/>
     {/*
       <div>
         <a href="https://vite.dev" target="_blank">
